+++ src/fastq.h
@@ -8,7 +8,6 @@
 #include "zran.h"
 #include "util.h"
 #include "sqlite3.h"
-#include "pycompat.h"
 
 typedef struct {
 	PyObject_HEAD