+++ src/identifier.h
@@ -3,7 +3,6 @@
 #include "Python.h"
 #include <stdint.h>
 #include "sqlite3.h"
-#include "pycompat.h"
 
 typedef struct {
 	PyObject_HEAD